requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
psycopg2-binary==2.9.3
python-magic==0.4.25
sqlparse==0.4.2
